Refactoring out of Jupyter notebook
Code now contained in a flat .py file (I intend to spin certain portions out into separate files, but this is a start).
Also very much refactored the code to both match PEP8 standard, as well as cleaning up some potential future errors.  Much more cleaning up to do though.
