+++ .idea/dictionaries/Daniel.xml
@@ -1,0 +1,41 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="Daniel">
+    <words>
+      <w>affordance</w>
+      <w>anonfirst</w>
+      <w>anonymized</w>
+      <w>dereferenced</w>
+      <w>deriv</w>
+      <w>derivationally</w>
+      <w>drspp</w>
+      <w>framenet</w>
+      <w>graphml</w>
+      <w>hacky</w>
+      <w>hardcode</w>
+      <w>hypernyms</w>
+      <w>hyponyms</w>
+      <w>ifline</w>
+      <w>indir</w>
+      <w>itemcount</w>
+      <w>networkx</w>
+      <w>nyms</w>
+      <w>objectnode</w>
+      <w>pred</w>
+      <w>predicateswitcher</w>
+      <w>propcount</w>
+      <w>psychomotor</w>
+      <w>qcheck</w>
+      <w>questionswitcher</w>
+      <w>refvarnodes</w>
+      <w>subjectnode</w>
+      <w>subjectref</w>
+      <w>substr</w>
+      <w>synset</w>
+      <w>synsets</w>
+      <w>tert</w>
+      <w>verbed</w>
+      <w>verbnet</w>
+      <w>wordnet</w>
+    </words>
+  </dictionary>
+</component>
