+++ python/jsonFile.txt
@@ -1,0 +1,1 @@
+{"directed": true, "multigraph": true, "graph": {}, "nodes": [{"value": "A", "id": "Item0"}, {"value": "Psychomotor-Vigilance", "id": "ItemName0"}, {"value": "", "id": "ItemAffordance0"}, {"value": "", "id": "ItemDescription0"}, {"value": "task", "id": "ItemRole0"}, {"value": "C", "id": "Item1"}, {"value": "Acknowledge", "id": "ItemName1"}, {"value": "", "id": "ItemAffordance1"}, {"value": "", "id": "ItemDescription1"}, {"value": "button", "id": "ItemRole1"}, {"value": "E", "id": "Item2"}, {"value": "", "id": "ItemName2"}, {"value": "", "id": "ItemAffordance2"}, {"value": "", "id": "ItemDescription2"}, {"value": "box", "id": "ItemRole2"}, {"value": "H", "id": "Item3"}, {"value": "", "id": "ItemName3"}, {"value": "", "id": "ItemAffordance3"}, {"value": "", "id": "ItemDescription3"}, {"value": "GROUP", "id": "ItemRole3"}, {"value": "G", "id": "Item4"}, {"value": "", "id": "ItemName4"}, {"value": "appear", "id": "ItemAffordance4"}, {"value": "", "id": "ItemDescription4"}, {"value": "target", "id": "ItemRole4"}, {"value": "I", "id": "Item5"}, {"value": "", "id": "ItemName5"}, {"value": "", "id": "ItemAffordance5"}, {"value": "", "id": "ItemDescription5"}, {"value": "letter", "id": "ItemRole5"}, {"value": "K", "id": "Action6"}, {"value": "appear", "id": "ActionVerb6"}, {"value": "Kmod", "id": "ModPP7"}, {"value": "in", "id": "ModPPPrep7"}, {"value": "L", "id": "Item8"}, {"value": "", "id": "ItemName8"}, {"value": "click|remember", "id": "ItemAffordance8"}, {"value": "", "id": "ItemDescription8"}, {"value": "subject", "id": "ItemRole8"}, {"value": "M", "id": "Action9"}, {"value": "click", "id": "ActionVerb9"}, {"value": "N", "id": "Action10"}, {"value": "remember", "id": "ActionVerb10"}, {"value": "O", "id": "Property11"}, {"value": "active", "id": "PropertyAdjective11"}, {"value": "", "id": "PropertySecondaryObject11"}, {"value": "", "id": "PropertyTertiaryObject11"}, {"value": "pos", "id": "PropertyDegree11"}, {"value": "", "id": "PropertyCompTarget11"}], "links": [{"value": "ItemHasName", "source": "Item0", "target": "ItemName0", "key": 0}, {"value": "ItemHasAffordance", "source": "Item0", "target": "ItemAffordance0", "key": 0}, {"value": "ItemHasDescription", "source": "Item0", "target": "ItemDescription0", "key": 0}, {"value": "ItemHasRole", "source": "Item0", "target": "ItemRole0", "key": 0}, {"value": "HasA", "source": "Item0", "target": "Item3", "key": 0}, {"value": "Is", "source": "Item0", "target": "Property11", "key": 0}, {"value": "ItemHasName", "source": "Item1", "target": "ItemName1", "key": 0}, {"value": "ItemHasAffordance", "source": "Item1", "target": "ItemAffordance1", "key": 0}, {"value": "ItemHasDescription", "source": "Item1", "target": "ItemDescription1", "key": 0}, {"value": "ItemHasRole", "source": "Item1", "target": "ItemRole1", "key": 0}, {"value": "IsTargetOf", "source": "Item1", "target": "Action9", "key": 0}, {"value": "ItemHasName", "source": "Item2", "target": "ItemName2", "key": 0}, {"value": "ItemHasAffordance", "source": "Item2", "target": "ItemAffordance2", "key": 0}, {"value": "ItemHasDescription", "source": "Item2", "target": "ItemDescription2", "key": 0}, {"value": "ItemHasRole", "source": "Item2", "target": "ItemRole2", "key": 0}, {"value": "IsMemberOf", "source": "Item2", "target": "Item3", "key": 0}, {"value": "isModifiedBy", "source": "Item2", "target": "ModPP7", "key": 0}, {"value": "ItemHasName", "source": "Item3", "target": "ItemName3", "key": 0}, {"value": "ItemHasAffordance", "source": "Item3", "target": "ItemAffordance3", "key": 0}, {"value": "ItemHasDescription", "source": "Item3", "target": "ItemDescription3", "key": 0}, {"value": "ItemHasRole", "source": "Item3", "target": "ItemRole3", "key": 0}, {"value": "HasMember", "source": "Item3", "target": "Item2", "key": 0}, {"value": "IsPartOf", "source": "Item3", "target": "Item0", "key": 0}, {"value": "HasMember", "source": "Item3", "target": "Item4", "key": 0}, {"value": "ItemHasName", "source": "Item4", "target": "ItemName4", "key": 0}, {"value": "ItemHasAffordance", "source": "Item4", "target": "ItemAffordance4", "key": 0}, {"value": "ItemHasDescription", "source": "Item4", "target": "ItemDescription4", "key": 0}, {"value": "ItemHasRole", "source": "Item4", "target": "ItemRole4", "key": 0}, {"value": "IsMemberOf", "source": "Item4", "target": "Item3", "key": 0}, {"value": "IsEquivalentTo", "source": "Item4", "target": "Item5", "key": 0}, {"value": "Performs", "source": "Item4", "target": "Action6", "key": 0}, {"value": "ItemHasName", "source": "Item5", "target": "ItemName5", "key": 0}, {"value": "ItemHasAffordance", "source": "Item5", "target": "ItemAffordance5", "key": 0}, {"value": "ItemHasDescription", "source": "Item5", "target": "ItemDescription5", "key": 0}, {"value": "ItemHasRole", "source": "Item5", "target": "ItemRole5", "key": 0}, {"value": "IsEquivalentTo", "source": "Item5", "target": "Item4", "key": 0}, {"value": "IsTargetOf", "source": "Item5", "target": "Action10", "key": 0}, {"value": "ActionHasVerb", "source": "Action6", "target": "ActionVerb6", "key": 0}, {"value": "IsPerformedBy", "source": "Action6", "target": "Item4", "key": 0}, {"value": "isModifiedBy", "source": "Action6", "target": "ModPP7", "key": 0}, {"value": "triggers", "source": "Action6", "target": "Action10", "key": 0}, {"value": "triggers", "source": "Action6", "target": "Action9", "key": 0}, {"value": "isTriggeredBy", "source": "Action6", "target": "Property11", "key": 0}, {"value": "ModPPHasPrep", "source": "ModPP7", "target": "ModPPPrep7", "key": 0}, {"value": "ModifiesVerb", "source": "ModPP7", "target": "Action6", "key": 0}, {"value": "modifiesObject", "source": "ModPP7", "target": "Item2", "key": 0}, {"value": "ItemHasName", "source": "Item8", "target": "ItemName8", "key": 0}, {"value": "ItemHasAffordance", "source": "Item8", "target": "ItemAffordance8", "key": 0}, {"value": "ItemHasDescription", "source": "Item8", "target": "ItemDescription8", "key": 0}, {"value": "ItemHasRole", "source": "Item8", "target": "ItemRole8", "key": 0}, {"value": "Performs", "source": "Item8", "target": "Action9", "key": 0}, {"value": "Performs", "source": "Item8", "target": "Action10", "key": 0}, {"value": "ActionHasVerb", "source": "Action9", "target": "ActionVerb9", "key": 0}, {"value": "IsPerformedBy", "source": "Action9", "target": "Item8", "key": 0}, {"value": "HasTarget", "source": "Action9", "target": "Item1", "key": 0}, {"value": "isTriggeredBy", "source": "Action9", "target": "Action6", "key": 0}, {"value": "ActionHasVerb", "source": "Action10", "target": "ActionVerb10", "key": 0}, {"value": "IsPerformedBy", "source": "Action10", "target": "Item8", "key": 0}, {"value": "HasTarget", "source": "Action10", "target": "Item5", "key": 0}, {"value": "isTriggeredBy", "source": "Action10", "target": "Action6", "key": 0}, {"value": "PropertyHasAdjective", "source": "Property11", "target": "PropertyAdjective11", "key": 0}, {"value": "PropertyHasSecondaryObject", "source": "Property11", "target": "PropertySecondaryObject11", "key": 0}, {"value": "PropertyHasTertiaryObject", "source": "Property11", "target": "PropertyTertiaryObject11", "key": 0}, {"value": "PropertyHasDegree", "source": "Property11", "target": "PropertyDegree11", "key": 0}, {"value": "PropertyHasCompTarget", "source": "Property11", "target": "PropertyCompTarget11", "key": 0}, {"value": "triggers", "source": "Property11", "target": "Action6", "key": 0}]}
